+++ src/lib/pepper/pepper.h
@@ -298,10 +298,10 @@ struct pepper_pointer_grab
 };
 
 PEPPER_API void
-pepper_pointer_set_position(pepper_pointer_t *pointer, int32_t x, int32_t y);
+pepper_pointer_set_position(pepper_pointer_t *pointer, double x, double y);
 
 PEPPER_API void
-pepper_pointer_get_position(pepper_pointer_t *pointer, int32_t *x, int32_t *y);
+pepper_pointer_get_position(pepper_pointer_t *pointer, double *x, double *y);
 
 PEPPER_API void
 pepper_pointer_set_focus(pepper_pointer_t *pointer, pepper_view_t *focus);
